chore: Move changelog into tracker [DHIS2-17714]
