+++ dhis-2/dhis-services/dhis-service-core/src/main/resources/org/hisp/dhis/trackedentityattributevalue/hibernate/TrackedEntityAttributeValueChangeLog.hbm.xml
@@ -4,7 +4,7 @@
         "http://www.hibernate.org/dtd/hibernate-mapping-3.0.dtd">
 
 <hibernate-mapping>
-    <class name="org.hisp.dhis.trackedentityattributevalue.TrackedEntityAttributeValueChangeLog"
+    <class name="org.hisp.dhis.tracker.changelog.TrackedEntityAttributeValueChangeLog"
            table="trackedentityattributevalueaudit">
 
         <id name="id" column="trackedentityattributevalueauditid">
