+++ dhis-2/dhis-services/dhis-service-tracker/src/main/java/org/hisp/dhis/tracker/changelog/DefaultTrackedEntityAttributeValueChangeLogService.java
@@ -25,7 +25,7 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package org.hisp.dhis.trackedentityattributevalue;
+package org.hisp.dhis.tracker.changelog;
 
 import java.util.List;
 import java.util.Set;
@@ -41,7 +41,7 @@ import org.springframework.stereotype.Service;
  * @author Morten Olav Hansen <mortenoh@gmail.com>
  */
 @RequiredArgsConstructor
-@Service("org.hisp.dhis.trackedentityattributevalue.TrackedEntityAttributeValueChangeLogService")
+@Service("org.hisp.dhis.tracker.changelog.TrackedEntityAttributeValueChangeLogService")
 public class DefaultTrackedEntityAttributeValueChangeLogService
     implements TrackedEntityAttributeValueChangeLogService {
   private final TrackedEntityAttributeValueChangeLogStore attributeValueChangeLogStore;
