+++ dhis-2/dhis-services/dhis-service-tracker/src/main/java/org/hisp/dhis/tracker/changelog/TrackedEntityAttributeValueChangeLogQueryParams.java
@@ -25,27 +25,34 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package org.hisp.dhis.trackedentityattributevalue;
+package org.hisp.dhis.tracker.changelog;
 
+import java.util.ArrayList;
 import java.util.List;
+import lombok.Data;
+import lombok.experimental.Accessors;
+import org.hisp.dhis.changelog.ChangeLogType;
+import org.hisp.dhis.common.Pager;
 import org.hisp.dhis.trackedentity.TrackedEntity;
+import org.hisp.dhis.trackedentity.TrackedEntityAttribute;
 
 /**
- * @author Morten Olav Hansen <mortenoh@gmail.com>
+ * Encapsulation of a web API request for tracked entity data value audit records.
+ *
+ * @author Lars Helge Overland
  */
-public interface TrackedEntityAttributeValueChangeLogService {
-  void addTrackedEntityAttributeValueChangLog(
-      TrackedEntityAttributeValueChangeLog trackedEntityAttributeValueChangeLog);
+@Data
+@Accessors(chain = true)
+public class TrackedEntityAttributeValueChangeLogQueryParams {
+  private List<TrackedEntityAttribute> trackedEntityAttributes = new ArrayList<>();
 
-  /**
-   * @deprecated use TrackedEntityChangeLogService.getEventChangeLog(UID) instead
-   */
-  @Deprecated(since = "2.41")
-  List<TrackedEntityAttributeValueChangeLog> getTrackedEntityAttributeValueChangeLogs(
-      TrackedEntityAttributeValueChangeLogQueryParams params);
+  private List<TrackedEntity> trackedEntities = new ArrayList<>();
 
-  int countTrackedEntityAttributeValueChangeLogs(
-      TrackedEntityAttributeValueChangeLogQueryParams params);
+  private List<ChangeLogType> auditTypes = new ArrayList<>();
 
-  void deleteTrackedEntityAttributeValueChangeLogs(TrackedEntity trackedEntity);
+  private Pager pager;
+
+  public boolean hasPager() {
+    return pager != null;
+  }
 }
